# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/libros": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Mostrar todos los libros",
                "operationId": "5de8af456fc24b6be3c98332057468d3",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Libros"
                ],
                "summary": "Añadir Usuario",
                "operationId": "e2798d77c0ea0393430855825b008022",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Titulo del libro nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Descripcion del titulo nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Añadir Libro y te devuelve el libro insertado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idLibro}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Mostrar un libro por su id",
                "operationId": "25fb622426857416567b80c11fc9ed45",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un libro por su id."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Libros"
                ],
                "summary": "Editar Libro",
                "operationId": "31a4c5a8dd8bd1e9c12d63b06fb3e847",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Nuevo titulo de libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Nueva descripcion del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Modificar un libro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idlibro} ": {
            "delete": {
                "tags": [
                    "Libros"
                ],
                "summary": "Eliminar Libro",
                "operationId": "6cc8870161c83ae57db742e35f1f3ef6",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Eliminar un libro por su Id y te muestra el libro borrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/libros/{idLibro}/usuarios": {
            "get": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Mostrar todos los usuarios que tienen determinado libro",
                "operationId": "a785f211cf6e49f2c42be9da866400a9",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios con ese libro."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/usuarios/{idusuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar un usuario por su id",
                "operationId": "49b788d6df98082222c8d8ab10537a98",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un usuario por su id."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Editar Usuario",
                "operationId": "4571026da8144edbc2cb4864b1352595",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nuevo nombre de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Nuevo email de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Modificar un usuario"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idusuario} ": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar Usuario",
                "operationId": "8981f2aca6f9889a32864dbc3d5c5920",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Eliminar un usuario por su Id y te muestra el usuario borrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idUsuario}/libros": {
            "get": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Mostrar todos los libros que tiene determinado usuario",
                "operationId": "4524370cd5ce286b9bb64f4c405abe02",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los libros de ese usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Prestamos"
                ],
                "summary": "Añadir un libro a determinado usuario",
                "operationId": "fe243cac67ccfc832153b7c183ec5271",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "libro_id",
                        "in": "query",
                        "description": "Id del libro prestado a usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios con libros."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}